project architecture changes
