+++ Travely.PropertyManager.API.Models/Travely.PropertyManager.API.Models.csproj
@@ -1,0 +1,21 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+  </PropertyGroup>
+
+	<ItemGroup>
+		<PackageReference Include="Google.Protobuf" Version="3.14.0" />
+		<PackageReference Include="Grpc" Version="2.34.1" />
+		<PackageReference Include="Grpc.Tools" Version="2.34.1" PrivateAssets="All">
+			<PrivateAssets>all</PrivateAssets>
+			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+		</PackageReference>
+	</ItemGroup>
+	
+	<ItemGroup>
+		<Protobuf Include="Protos\properties.proto" GrpcServices="Server" />
+	</ItemGroup>
+
+
+</Project>
