+++ Travely.PropertyManager.API/Properties/launchSettings.json
@@ -1,0 +1,18 @@
+﻿{
+    "profiles": {
+        "Travely.PropertyManager.API": {
+            "commandName": "Project",
+            "environmentVariables": {
+                "ASPNETCORE_ENVIRONMENT": "Development"
+            },
+            "dotnetRunMessages": "true",
+            "applicationUrl": "http://localhost:5100;https://localhost:5101"
+        },
+        "Docker": {
+            "commandName": "Docker",
+            "launchUrl": "{Scheme}://{ServiceHost}:{ServicePort}",
+            "publishAllPorts": true,
+            "useSSL": true
+        }
+    }
+}
