+++ Travely.PropertyManager.API/appsettings.json
@@ -1,0 +1,18 @@
+{
+    "Logging": {
+        "LogLevel": {
+            "Default": "Information",
+            "Microsoft": "Warning",
+            "Microsoft.Hosting.Lifetime": "Information"
+        }
+    },
+    "AllowedHosts": "*",
+    "Kestrel": {
+        "EndpointDefaults": {
+            "Protocols": "Http2"
+        }
+    },
+    "ConnectionStrings": {
+        "PropertyDbConnection": "Server=localhost\\SQLEXPRESS; Database=Property; Trusted_Connection=True;"
+    }
+}
