+++ Travely.PropertyManager.Data/EntityFramework/PropertyDbContext.cs
@@ -1,0 +1,23 @@
+﻿using System.Reflection;
+using Microsoft.EntityFrameworkCore;
+using Travely.PropertyManager.Data.Models;
+
+namespace Travely.PropertyManager.Data.EntityFramework
+{
+    public class PropertyDbContext : DbContext
+    {
+        public PropertyDbContext(DbContextOptions<PropertyDbContext> options)
+            : base(options)
+        {
+        }
+
+        public DbSet<Property> Properties { get; set; }
+
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            base.OnModelCreating(modelBuilder);
+
+            modelBuilder.ApplyConfigurationsFromAssembly(Assembly.GetAssembly(GetType()));
+        }
+    }
+}
