+++ Travely.PropertyManager.Service/Contracts/IPropertyService.cs
@@ -1,0 +1,15 @@
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
+using Travely.PropertyManager.Service.Models.Commands;
+using Travely.PropertyManager.Service.Models.Queries;
+using Travely.PropertyManager.Service.Models.Responses;
+
+namespace Travely.PropertyManager.Service.Contracts
+{
+    public interface IPropertyService
+    {
+        Task<int> AddAsync(AddPropertyCommand command);
+
+        Task<IEnumerable<PropertyResponse>> GetAsync(GetPropertiesQuery query);
+    }
+}
