+++ Travely.PropertyManager.Service/Models/Base/FilteringBaseModel.cs
@@ -1,0 +1,16 @@
+﻿namespace Travely.PropertyManager.Service.Models.Base
+{
+    public class FilteringBaseModel
+    {
+        public string FieldName { get; set; }
+        public string Value { get; set; }
+        public FilteringOperationType Type { get; set; }
+    }
+
+
+    public enum FilteringOperationType
+    {
+        Equals = 1,
+        Contains = 2
+    }
+}
