+++ Travely.PropertyManager.Service/Models/Base/OrderingBaseModel.cs
@@ -1,0 +1,8 @@
+﻿namespace Travely.PropertyManager.Service.Models.Base
+{
+    public class OrderingBaseModel
+    {
+        public string FieldName { get; set; }
+        public bool IsDescending { get; set; }
+    }
+}
