--- src/PropertyManager/Domain/Travely.PropertyManager.Domain/Models/Queries/GetPropertiesQuery.cs
@@ -1,12 +1,0 @@
-﻿using System.Collections.Generic;
-using Travely.PropertyManager.Domain.Entities.Base;
-
-namespace Travely.PropertyManager.Domain.Models.Queries
-{
-    public class GetPropertiesQuery
-    {
-        public ICollection<FilteringBaseModel> Filters { get; set; }
-        public ICollection<OrderingBaseModel> Orderings { get; set; }
-    }
-
-}
