--- src/PropertyManager/Infrastructure/Travely.PropertyManager.GrpcService/Dockerfile
@@ -1,22 +1,0 @@
-#See https://aka.ms/containerfastmode to understand how Visual Studio uses this Dockerfile to build your images for faster debugging.
-
-FROM mcr.microsoft.com/dotnet/aspnet:5.0-buster-slim AS base
-WORKDIR /app
-EXPOSE 80
-EXPOSE 443
-
-FROM mcr.microsoft.com/dotnet/sdk:5.0-buster-slim AS build
-WORKDIR /src
-COPY ["src/PropertyManager/Travely.PropertyManager.GrpcService/Travely.PropertyManager.GrpcService.csproj", "src/PropertyManager/Travely.PropertyManager.GrpcService/"]
-RUN dotnet restore "src/PropertyManager/Travely.PropertyManager.GrpcService/Travely.PropertyManager.GrpcService.csproj"
-COPY . .
-WORKDIR "/src/src/PropertyManager/Travely.PropertyManager.GrpcService"
-RUN dotnet build "Travely.PropertyManager.GrpcService.csproj" -c Release -o /app/build
-
-FROM build AS publish
-RUN dotnet publish "Travely.PropertyManager.GrpcService.csproj" -c Release -o /app/publish
-
-FROM base AS final
-WORKDIR /app
-COPY --from=publish /app/publish .
-ENTRYPOINT ["dotnet", "Travely.PropertyManager.GrpcService.dll"]
