--- src/PropertyManager/Shared/Travely.PropertyManager.Bootstrapper/Bootstrapper.AutoMapper.cs
@@ -1,14 +1,0 @@
-﻿using AutoMapper;
-using Microsoft.Extensions.DependencyInjection;
-using Travely.PropertyManager.Domain.MappingProfiles;
-
-namespace Travely.PropertyManager.Bootstrapper
-{
-    public static partial class Bootstrapper
-    {
-        public static void ConfigureAutoMapper(IServiceCollection services)
-        {
-            services.AddAutoMapper(typeof(PropertyMappingProfile));
-        }
-    }
-}
