--- src/PropertyManager/Shared/Travely.PropertyManager.Bootstrapper/Bootstrapper.Domain.cs
@@ -1,14 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-using Travely.PropertyManager.Domain.Contracts;
-using Travely.PropertyManager.Domain.Services;
-
-namespace Travely.PropertyManager.Bootstrapper
-{
-    public static partial class Bootstrapper
-    {
-        public static void RegisterServices(IServiceCollection services)
-        {
-            services.AddTransient<IPropertyService, PropertyService>();
-        }
-    }
-}
